dashboard/Crucible/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
google-genai
itsdangerous
python-multipart
